requirements.txt
sqlalchemy~=1.3.16
gino~=1.0.0
python-dotenv
aiogram~=2.8
config~=0.5.0
pyenchant~=3.0.1